Work on waste map structure and rename namespace to BinPacking
